+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 -   Don't replace selection on comment insertion.
 -   Don't focus on tables on cell click.
 -   Fix off by one on table row and column deletion.
+-   More context sensitive paste.
 
 # 0.5.7 - 2023-11-18